Plan the exported C wrappers as artifacts
The `extern "C"` wrapper for a declared `#[prebindgen]` function was built
inside `Prebindgen::on_function`, which the writer calls per item while
assembling the file. Everything the wrapper needed — the exported symbol, the
path it calls, whether `.panic()` was declared, which parameters name an opaque
resource the alias preflight compares — was read off the declaration object at
that moment, so the writer's item walk was what decided when a wrapper came
into existence.

The wrapper is now planned once, when the C generation plan is frozen, and
joins the assembly as a `CFinalArtifact::Wrapper`. `CWrapper` holds its own
frozen state and an `Rc` to the frozen plan its boundary sites live in;
rendering reads nothing else. That a `Result`-returning function's error type
was declared with `.error()` is checked while planning rather than while
rendering, which is the one check that moves.

`CbindgenBuilder::on_function` returns nothing for now. The method itself goes
when JniGen's wrappers follow in the next step.

`ordinary_wrapper_rendering_cannot_resume_legacy_planning` follows the renderer
to its new home and keeps naming what the body must and must not reach, and a
new fence asserts the planned wrapper retains no declaration object, registry
or compilation cache — only the frozen plan.

Generated Rust, C headers and Kotlin are byte-for-byte unchanged.

Part of #581.
