+++ prebindgen-c/src/assembly.rs
@@ -1,0 +1,536 @@
+//! What the generated C-facing Rust file is assembled from.
+//!
+//! Each final artifact is planned when the C generation plan is frozen and
+//! rendered at the final writing boundary, so nothing it needs is looked up in
+//! a live registry while the file is assembled. A wrapper — the exported
+//! `extern "C"` function one declared `#[prebindgen]` function becomes — reads
+//! its own frozen state plus the boundary sites of its own function, which the
+//! same frozen plan holds.
+
+use std::rc::Rc;
+
+use prebindgen_registry::{
+    generation::GenerationPlan,
+    write::{ArtifactKey, RustArtifact},
+};
+
+use super::*;
+
+/// One final artifact of the generated Rust file.
+pub(crate) enum CFinalArtifact {
+    /// A private converter, carrying one value across the boundary.
+    Converter(Box<crate::chain::CFunction>),
+    /// The exported wrapper for one declared `#[prebindgen]` function.
+    Wrapper(Box<CWrapper>),
+}
+
+impl RustArtifact for CFinalArtifact {
+    fn key(&self) -> ArtifactKey {
+        match self {
+            Self::Converter(converter) => converter.key(),
+            Self::Wrapper(wrapper) => wrapper.key(),
+        }
+    }
+
+    fn render(&self, emit: &prebindgen_registry::RustWriter) -> Vec<syn::Item> {
+        match self {
+            Self::Converter(converter) => converter.render(emit),
+            Self::Wrapper(wrapper) => wrapper.render(emit),
+        }
+    }
+}
+
+/// One declared function's exported wrapper, frozen at the end of resolution.
+pub(crate) struct CWrapper {
+    /// The plan the wrapper's own boundary sites are read from.
+    generation: Rc<GenerationPlan<crate::compile::CRepresentation>>,
+    /// The `#[prebindgen]` function this wrapper exports.
+    function: prebindgen_registry::flat::Function,
+    /// The `#[no_mangle]` symbol the wrapper is exported under.
+    symbol: syn::Ident,
+    /// Path the wrapper calls the source function by.
+    call_path: syn::Path,
+    /// Whether `.panic()` was declared, which is what allows a fallible
+    /// conversion in a function that returns no `Result`.
+    panic: bool,
+    /// Per parameter, the resource its argument names and how the call uses
+    /// it — `None` for a parameter that names no single owned resource. The
+    /// alias preflight compares these; declaring which types are opaque is a
+    /// binding question, so it is answered while planning.
+    alias_slots: Vec<Option<(TypeKey, AliasAccess)>>,
+}
+
+impl CWrapper {
+    /// Plan the wrapper for one declared function.
+    pub(crate) fn new(
+        decls: &CbindgenBuilder,
+        generation: &Rc<GenerationPlan<crate::compile::CRepresentation>>,
+        function: &prebindgen_registry::flat::Function,
+    ) -> Self {
+        let name = &function.name;
+        if let Some((_, err)) = function.ret.fallible_parts() {
+            let err_key = err.key();
+            assert!(
+                decls.error.contains(&err_key),
+                "Cbindgen: function `{}` returns `Result<_, {}>` but `{}` is not a \
+                 declared error type — add `.data_struct({}).error()`",
+                name,
+                err_key,
+                err_key,
+                err_key,
+            );
+        }
+        Self {
+            generation: Rc::clone(generation),
+            function: function.clone(),
+            symbol: decls.fn_symbol(name),
+            call_path: decls.src_fn(name),
+            panic: decls.functions.get(name).is_some_and(|cfg| cfg.panic),
+            alias_slots: function
+                .params
+                .iter()
+                .map(|param| decls.alias_slot_of(&param.ty))
+                .collect(),
+        }
+    }
+
+    /// One boundary site of this wrapper's own function.
+    fn site(
+        &self,
+        role: prebindgen_registry::recipe::Role,
+    ) -> &prebindgen_registry::generation::SitePlan<crate::compile::CRepresentation> {
+        let id = prebindgen_registry::generation::SiteId::new(prebindgen_registry::recipe::Site {
+            owner: self.function.name.clone(),
+            role,
+        });
+        let site = self
+            .generation
+            .site(&id)
+            .unwrap_or_else(|| panic!("C generation plan has no site {id}"));
+        assert!(
+            matches!(
+                site.cleanup(),
+                prebindgen_registry::generation::Cleanup::None
+            ),
+            "ordinary C sites cannot carry a deferred cleanup operation"
+        );
+        site
+    }
+
+    /// Assemble this wrapper's `#[no_mangle] extern "C"` function.
+    fn render_fn(&self, emit: &prebindgen_registry::RustWriter) -> syn::ItemFn {
+        let f = &self.function;
+        let orig = &f.name;
+        let call_path = &self.call_path;
+        let sym = &self.symbol;
+
+        // The ELEMENT: a signature is a parameter list and a return, both
+        // already classified. An elided return is `TypeKind::Unit`, which is
+        // the `ReturnType::Default` arm this used to write.
+
+        let has_fallible_input = f.params.iter().enumerate().any(|(index, _)| {
+            self.site(prebindgen_registry::recipe::Role::Param { index })
+                .abi()
+                .payload()
+                .failure()
+                == prebindgen_registry::generation::Failure::Fallible
+        });
+
+        // Peel an outer `Result<_, E>`; `value_ty` is the success/return value.
+        // Off `TypeKind::Fallible`, where `result_parts` found the `Result` in a
+        // path first — and both sides come back as readings, so everything
+        // downstream of here reads too.
+        let (value_ty, err_reading) = match f.ret.fallible_parts() {
+            Some((ok, e)) => (ok, Some(e)),
+            None => (&f.ret, None),
+        };
+        let value_site = (!matches!(value_ty.kind(), TypeKind::Unit))
+            .then(|| self.site(prebindgen_registry::recipe::Role::Return));
+        let has_fallible_output = value_site.is_some_and(|site| {
+            site.abi().payload().failure() == prebindgen_registry::generation::Failure::Fallible
+        });
+
+        // Error wiring: that the error type is declared via `.error()` was
+        // checked when the wrapper was planned.
+        let err_bits = err_reading.map(|err| {
+            let site = self.site(prebindgen_registry::recipe::Role::Error);
+            let crate::compile::CValue::Direct {
+                wire, converter, ..
+            } = site.abi().payload()
+            else {
+                panic!("C error site must have one direct wire");
+            };
+            (
+                wire.clone(),
+                converter.ident(emit),
+                emit.emit_source_type(err),
+            )
+        });
+
+        // No `Result` channel ⇒ a fallible input must be declared `.panic()`.
+        if err_reading.is_none() {
+            assert!(
+                !(has_fallible_input || has_fallible_output) || self.panic,
+                "Cbindgen: function `{}` has a fallible binding conversion but does not \
+                 return `Result`; add \
+                 `.panic()` after its `.function(...)` declaration to allow aborting \
+                 on the internal error, or change its signature",
+                orig,
+            );
+        }
+
+        // Structural lowering of the (present/ok) value, then the null-niche rule:
+        //   * Result + a free pointer niche  → NULL marks `Err` (value in-band);
+        //   * Result without a free niche     → `bool` status, value to out-params;
+        //   * no Result                       → field 0 is the C return, rest out.
+        let shape = value_site.map_or(
+            FrozenValueLayout {
+                fields: Vec::new(),
+                niches: Niches::empty(),
+            },
+            |site| FrozenValueLayout {
+                fields: site.abi().payload().fields(),
+                niches: site.abi().payload().effective_niches(),
+            },
+        );
+        let result_slot = shape.niches.clone().carve().map(|(slot, _)| slot);
+        let result_in_band = err_reading.is_some() && result_slot.is_some();
+        let field0_is_return = result_in_band || err_reading.is_none();
+
+        // Partition fields into the (optional) C return value + out-parameters,
+        // and pick C names for the out-params (see `out_param_name`).
+        let mut targets: Vec<TokenStream> = Vec::new();
+        let mut out_fields: Vec<&WireField> = Vec::new();
+        // `field0_wire` is the wire of the value's primary field when that field
+        // is carried by the C return slot (modes A/D); `None` for mode B and unit.
+        let field0_wire: Option<syn::Type> = if field0_is_return {
+            shape.fields.first().map(|f| f.wire.clone())
+        } else {
+            None
+        };
+        if field0_is_return {
+            if !shape.fields.is_empty() {
+                targets.push(quote!(__ret));
+                out_fields.extend(shape.fields[1..].iter());
+            }
+        } else {
+            out_fields.extend(shape.fields.iter());
+        }
+        let prefixed = out_fields.iter().any(|wf| wf.suffix.is_empty());
+        let out_names: Vec<syn::Ident> = out_fields
+            .iter()
+            .map(|wf| out_param_name(wf.suffix, prefixed))
+            .collect();
+        for name in &out_names {
+            targets.push(quote!(*#name));
+        }
+        let out_param_decls: Vec<TokenStream> = out_fields
+            .iter()
+            .zip(&out_names)
+            .map(|(wf, name)| {
+                let wire = &wf.wire;
+                quote!(#name: *mut #wire)
+            })
+            .collect();
+
+        // C wrapper return type: the payload's field 0 (modes A/D), `bool` status
+        // (mode B), or `void` (a unit value with no `Result`).
+        let c_return: Option<syn::Type> = if field0_is_return {
+            field0_wire.clone()
+        } else {
+            Some(syn::parse_quote!(bool))
+        };
+
+        // Input decode: route a fallible-input failure to the error out-param
+        // (with the wrapper's fail value) when there is a `Result`, else panic.
+        let fail_return = if result_in_band {
+            let slot = result_slot.as_ref().expect("in-band result has a niche");
+            let value = &slot.value;
+            quote!(#value)
+        } else {
+            quote!(false)
+        };
+        let mut planned_routes = f
+            .params
+            .iter()
+            .enumerate()
+            .filter_map(|(index, _)| {
+                self.site(prebindgen_registry::recipe::Role::Param { index })
+                    .failure_route()
+            })
+            .chain(
+                value_site
+                    .into_iter()
+                    .filter_map(|site| site.failure_route()),
+            );
+        let planned_route = planned_routes.next();
+        assert!(
+            planned_routes.all(|route| Some(route) == planned_route),
+            "C function sites disagree on their frozen failure route"
+        );
+        match planned_route {
+            Some(crate::compile::CFailureRoute::Panic) => {
+                assert!(
+                    err_bits.is_none(),
+                    "panic route attached to a Result function"
+                );
+            }
+            Some(crate::compile::CFailureRoute::Error(error_site)) => {
+                assert!(
+                    err_bits.is_some(),
+                    "error route attached without a Result channel"
+                );
+                let expected = prebindgen_registry::generation::SiteId::new(
+                    prebindgen_registry::recipe::Site {
+                        owner: orig.clone(),
+                        role: prebindgen_registry::recipe::Role::Error,
+                    },
+                );
+                assert_eq!(
+                    error_site, &expected,
+                    "failure route names the wrong error site"
+                );
+            }
+            None => {}
+        }
+        let input_route = match &err_bits {
+            Some((_, e_conv, e_ty_src)) => ErrRoute::Result {
+                e_conv,
+                e_ty_src: e_ty_src.clone(),
+                fail_return: fail_return.clone(),
+            },
+            None => ErrRoute::Panic,
+        };
+        let (in_params, decodes, call_args) = self.planned_inputs(&input_route, emit);
+        let call = quote!(#call_path(#(#call_args),*));
+
+        let e_param = err_bits
+            .as_ref()
+            .map(|(err_wire, _, _)| quote!(e: *mut #err_wire));
+        let ret_arrow = c_return.as_ref().map(|w| quote!(-> #w));
+
+        // Assemble the body per the three structural modes.
+        let body = match (&err_bits, field0_is_return) {
+            // No `Result`: straight-line. `void` when there are no fields.
+            (None, _) => {
+                if let Some(field0_wire) = field0_wire.as_ref() {
+                    let enc = value_site.map_or_else(TokenStream::new, |site| {
+                        site.abi()
+                            .payload()
+                            .encode(quote!(__v), &targets, &input_route, emit)
+                    });
+                    quote!(
+                        #(#decodes)*
+                        let __v = #call;
+                        let __ret: #field0_wire;
+                        #enc
+                        __ret
+                    )
+                } else {
+                    quote!( #(#decodes)* #call; )
+                }
+            }
+            // `Result` with a free niche: value in-band, NULL marks `Err`.
+            (Some((_, e_conv, _)), true) => {
+                let field0_wire = field0_wire.as_ref().expect("in-band ⇒ pointer return");
+                let null = &result_slot
+                    .as_ref()
+                    .expect("in-band result has a niche")
+                    .value;
+                let enc = value_site.map_or_else(TokenStream::new, |site| {
+                    site.abi()
+                        .payload()
+                        .encode(quote!(__v), &targets, &input_route, emit)
+                });
+                quote!(
+                    #(#decodes)*
+                    match #call {
+                        ::core::result::Result::Ok(__v) => { let __ret: #field0_wire; #enc __ret }
+                        ::core::result::Result::Err(__err) => {
+                            if !e.is_null() { *e = #e_conv(__err); }
+                            #null
+                        }
+                    }
+                )
+            }
+            // `Result` without a free niche: `bool` status, value to out-params.
+            (Some((_, e_conv, _)), false) => {
+                let enc = value_site.map_or_else(TokenStream::new, |site| {
+                    site.abi()
+                        .payload()
+                        .encode(quote!(__v), &targets, &input_route, emit)
+                });
+                quote!(
+                    #(#decodes)*
+                    match #call {
+                        ::core::result::Result::Ok(__v) => { #enc true }
+                        ::core::result::Result::Err(__err) => {
+                            if !e.is_null() { *e = #e_conv(__err); }
+                            false
+                        }
+                    }
+                )
+            }
+        };
+
+        syn::parse_quote! {
+            #[no_mangle]
+            #[allow(non_snake_case, unused_mut, unused_variables, unused_unsafe, dead_code)]
+            pub unsafe extern "C" fn #sym(
+                #(#in_params,)*
+                #(#out_param_decls,)*
+                #e_param
+            ) #ret_arrow {
+                #body
+            }
+        }
+    }
+
+    /// The runtime **alias preflight**: reject a call whose arguments name the
+    /// same resource in a combination that would reconstruct or invalidate it
+    /// twice, *before* any conversion runs.
+    ///
+    /// `z_combine(primary: ZThing, fallback: ZThing)` is a supported
+    /// declaration; called as `z_combine(x, x)` it reaches `Box::from_raw`
+    /// twice on one allocation. So is `f(a: ZThing, b: &ZThing)` — the borrow
+    /// dangles the moment the consume takes ownership — and `f(a: &mut T, b:
+    /// &T)`, where the exclusive reference is not exclusive at all. Rejecting
+    /// the *declaration* is not an option: these are shapes that ship today, so
+    /// removing them would be a regression that a later stage would have to
+    /// undo.
+    ///
+    /// Emitted whenever a call has **at least one `Consume` or
+    /// `ExclusiveBorrow`** and **any other active access in the same resource
+    /// domain**. Stated that way rather than as "two or more consumed
+    /// parameters", which would skip exactly the two mixed cases above.
+    ///
+    /// A NULL pointer names no resource — two NULLs are not an alias — and is
+    /// rejected by each converter's own null check, so the comparison skips it.
+    ///
+    /// Shared/shared is *not* rejected: two `&T` to one resource is legal Rust
+    /// and legal C.
+    fn alias_preflight(&self, route: &ErrRoute) -> Option<TokenStream> {
+        let slots: Vec<(syn::Ident, TypeKey, AliasAccess)> = self
+            .function
+            .params
+            .iter()
+            .zip(&self.alias_slots)
+            .filter_map(|(param, slot)| {
+                slot.as_ref()
+                    .map(|(key, access)| (param.name.clone(), key.clone(), *access))
+            })
+            .collect();
+
+        let on_err = route_message(route);
+        let mut checks: Vec<TokenStream> = Vec::new();
+        for i in 0..slots.len() {
+            for j in (i + 1)..slots.len() {
+                let (a, a_key, a_access) = &slots[i];
+                let (b, b_key, b_access) = &slots[j];
+                if a_key != b_key {
+                    continue;
+                }
+                // At least one side must be exclusive about the resource.
+                if matches!(a_access, AliasAccess::Shared)
+                    && matches!(b_access, AliasAccess::Shared)
+                {
+                    continue;
+                }
+                let msg = format!(
+                    "aliasing arguments: `{a}` ({}) and `{b}` ({}) are the same `{}` — a \
+                     consumed or exclusively-borrowed resource may not be named twice in one call",
+                    a_access.describe(),
+                    b_access.describe(),
+                    a_key.as_str(),
+                );
+                checks.push(quote!(
+                    if !(#a as *const ()).is_null() && (#a as *const ()) == (#b as *const ()) {
+                        let __msg = ::std::string::String::from(#msg);
+                        #on_err
+                    }
+                ));
+            }
+        }
+        (!checks.is_empty()).then(|| quote!(#(#checks)*))
+    }
+
+    /// Build the wire param list, per-input decode statements, and call-site
+    /// argument expressions. Fallible inputs (converter returns `Result<_,
+    /// String>`) route their `Err(msg)` per `route`; infallible inputs decode
+    /// directly.
+    fn planned_inputs(
+        &self,
+        route: &ErrRoute<'_>,
+        emit: &prebindgen_registry::RustWriter,
+    ) -> (Vec<TokenStream>, Vec<TokenStream>, Vec<TokenStream>) {
+        let mut params = Vec::new();
+        let mut decodes: Vec<TokenStream> = self.alias_preflight(route).into_iter().collect();
+        let mut call_args = Vec::new();
+
+        for (index, param) in self.function.params.iter().enumerate() {
+            let ident = &param.name;
+            let site = self.site(prebindgen_registry::recipe::Role::Param { index });
+            match site.abi().payload() {
+                crate::compile::CValue::BorrowedInput {
+                    element,
+                    wire,
+                    reinterpret,
+                } => {
+                    let len_id = format_ident!("{}_len", ident);
+                    let source = emit.emit_source_type(element);
+                    params.push(quote!(#ident: #wire));
+                    params.push(quote!(#len_id: usize));
+                    let from_parts = if *reinterpret {
+                        quote!(::core::slice::from_raw_parts(
+                            #ident as *const #source,
+                            #len_id,
+                        ))
+                    } else {
+                        quote!(::core::slice::from_raw_parts(#ident, #len_id))
+                    };
+                    decodes.push(quote!(
+                        let #ident: &[#source] = if #ident.is_null() {
+                            &[]
+                        } else {
+                            #from_parts
+                        };
+                    ));
+                }
+                crate::compile::CValue::Direct {
+                    wire, converter, ..
+                } => {
+                    let conv = converter.ident(emit);
+                    params.push(quote!(#ident: #wire));
+                    if converter.fallible() {
+                        let on_err = route_message(route);
+                        decodes.push(quote!(
+                            let #ident = match #conv(#ident) {
+                                ::core::result::Result::Ok(__v) => __v,
+                                ::core::result::Result::Err(__msg) => { #on_err }
+                            };
+                        ));
+                    } else {
+                        decodes.push(quote!(let #ident = #conv(#ident);));
+                    }
+                }
+                _ => panic!("C input site has an output-only ABI plan"),
+            }
+            call_args.push(quote!(#ident));
+        }
+
+        (params, decodes, call_args)
+    }
+}
+
+impl RustArtifact for CWrapper {
+    fn key(&self) -> ArtifactKey {
+        ArtifactKey::Artifact(
+            prebindgen_registry::generation::ArtifactId::new("c-wrapper", self.symbol.to_string())
+                .expect("an exported symbol is a non-empty artifact name"),
+        )
+    }
+
+    fn render(&self, emit: &prebindgen_registry::RustWriter) -> Vec<syn::Item> {
+        vec![syn::Item::Fn(self.render_fn(emit))]
+    }
+}
